Revert WP-CLI caching
