+++ packages/cli/package.json
@@ -33,7 +33,6 @@
     "@wp-tester/config": "*",
     "@wp-tester/phpunit": "*",
     "@wp-tester/results": "*",
-    "@wp-tester/runtime": "*",
     "@wp-tester/smoke-tests": "*",
     "ajv": "^8.17.1",
     "picocolors": "^1.1.1",
