+++ packages/runtime/src/index.ts
@@ -21,11 +21,6 @@ export { startPlayground, stopPlayground, defaultWpCliPath } from "./playground.
 
 // Export WP CLI utilities
 export { wpCli, getWordPressVersion } from "./wp-cli.js";
-export { downloadWpCli, WP_CLI_VFS_PATH } from "./wp-cli-cache.js";
-
-// Export caching utilities
-export { cacheFetch, CACHE_FOREVER, CACHE_DISABLED, CACHE_1_HOUR, CACHE_1_DAY, CACHE_1_WEEK } from "./cache-fetch.js";
-export type { CacheFetchOptions } from "./cache-fetch.js";
 
 // Export HTTP utilities
 export { request } from "./request.js";
